test: replace Tavus mock transport script with mock server
Rework the manual Tavus testing script into a standalone mock server
that supports the conversation.echo app message flow used by the new
audio delivery path.
